deps: update V8 to 8.9.255.19
PR-URL: https://github.com/nodejs/node/pull/37330

+++ deps/v8/src/compiler/scheduler.cc
@@ -478,14 +478,14 @@ class CFGBuilder : public ZoneObject {
 
     BranchHint hint_from_profile = BranchHint::kNone;
     if (const ProfileDataFromFile* profile_data = scheduler_->profile_data()) {
-      uint32_t block_zero_count =
+      double block_zero_count =
           profile_data->GetCounter(successor_blocks[0]->id().ToSize());
-      uint32_t block_one_count =
+      double block_one_count =
           profile_data->GetCounter(successor_blocks[1]->id().ToSize());
       // If a branch is visited a non-trivial number of times and substantially
       // more often than its alternative, then mark it as likely.
-      constexpr uint32_t kMinimumCount = 100000;
-      constexpr uint32_t kThresholdRatio = 4000;
+      constexpr double kMinimumCount = 100000;
+      constexpr double kThresholdRatio = 4000;
       if (block_zero_count > kMinimumCount &&
           block_zero_count / kThresholdRatio > block_one_count) {
         hint_from_profile = BranchHint::kTrue;
